+++ blog/src/models/mod.rs
@@ -1,0 +1,10 @@
+pub enum PublishState {
+    Draft,
+    Public,
+    Unlisted,
+    Archived,
+}
+
+pub use self::blog_post::BlogPost;
+mod blog_post;
+pub mod dto;
